Add files via upload
The application you will build is intended for gathering survey data on people's lifestyle
preferences. The app should be a desktop, web, or mobile application that can be
placed in a public area for individuals to fill out a survey. The objective is to gather as
many responses as possible for data analysis purposes. Please feel free to contact us
for clarification on these requirements.